requirements.txt
aiobotocore==2.5.4
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.3.0
attrs==23.2.0
blinker==1.8.1
botocore==1.31.17
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
fsspec==2023.9.2
gitdb==4.0.11
gitpython==3.1.43
idna==3.7
jinja2==3.1.3
jmespath==1.0.1
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
protobuf==4.25.3
pyarrow==16.0.0
pydeck==0.8.0
pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxet==0.1.9
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3fs==2023.9.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
streamlit==1.33.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
